--- .pipelines/templates/template-clean-e2e-deps.yml
@@ -1,10 +1,0 @@
-steps:
-- script: |
-    . secrets/env
-
-    set -x
-    . ./hack/e2e/run-rp-and-e2e.sh
-
-    clean_e2e
-  displayName: 🧹 Clean cluster RG, and Vnet
-  condition: always()
