+++ hack/kubeadminkubeconfig/kubeadminkubeconfig.go
@@ -5,28 +5,30 @@ package main
 
 import (
 	"context"
+	"encoding/json"
 	"os"
 
-	"github.com/Azure/go-autorest/autorest/azure"
 	"github.com/Azure/go-autorest/autorest/azure/auth"
 	"github.com/sirupsen/logrus"
 
 	utillog "github.com/Azure/ARO-RP/pkg/util/log"
-	"github.com/Azure/ARO-RP/test/util/kubeconfig"
+	"github.com/Azure/ARO-RP/test/util/kubeadminkubeconfig"
 )
 
 func run(ctx context.Context, log *logrus.Entry, resourceID string) error {
-	res, err := azure.ParseResourceID(resourceID)
-	if err != nil {
-		return err
-	}
-
 	authorizer, err := auth.NewAuthorizerFromEnvironment()
 	if err != nil {
 		return err
 	}
 
-	return kubeconfig.NewManager(log, res.SubscriptionID, authorizer).Print(ctx, res.ResourceGroup, res.ResourceName)
+	adminKubeconfig, err := kubeadminkubeconfig.Get(ctx, log, authorizer, resourceID)
+	if err != nil {
+		return err
+	}
+
+	e := json.NewEncoder(os.Stdout)
+	e.SetIndent("", "    ")
+	return e.Encode(adminKubeconfig)
 }
 
 func main() {
