+++ pkg/util/azureclient/graphrbac/applications.go
@@ -12,7 +12,6 @@ import (
 
 // ApplicationsClient is a minimal interface for azure ApplicationsClient
 type ApplicationsClient interface {
-	ApplicationsClientAddons
 	Create(ctx context.Context, parameters graphrbac.ApplicationCreateParameters) (result graphrbac.Application, err error)
 	GetServicePrincipalsIDByAppID(ctx context.Context, applicationID string) (result graphrbac.ServicePrincipalObjectResult, err error)
 	Delete(ctx context.Context, applicationObjectID string) (result autorest.Response, err error)
