+++ pkg/util/mocks/azureclient/graphrbac/graphrbac.go
@@ -81,21 +81,6 @@ func (mr *MockApplicationsClientMockRecorder) GetServicePrincipalsIDByAppID(arg0
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetServicePrincipalsIDByAppID", reflect.TypeOf((*MockApplicationsClient)(nil).GetServicePrincipalsIDByAppID), arg0, arg1)
 }
 
-// List mocks base method
-func (m *MockApplicationsClient) List(arg0 context.Context, arg1 string) ([]graphrbac.Application, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "List", arg0, arg1)
-	ret0, _ := ret[0].([]graphrbac.Application)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// List indicates an expected call of List
-func (mr *MockApplicationsClientMockRecorder) List(arg0, arg1 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "List", reflect.TypeOf((*MockApplicationsClient)(nil).List), arg0, arg1)
-}
-
 // MockServicePrincipalClient is a mock of ServicePrincipalClient interface
 type MockServicePrincipalClient struct {
 	ctrl     *gomock.Controller
