--- test/util/kubeconfig/kubeconfig.go
@@ -1,49 +1,0 @@
-package kubeconfig
-
-// Copyright (c) Microsoft Corporation.
-// Licensed under the Apache License 2.0.
-
-import (
-	"strings"
-
-	v1 "k8s.io/client-go/tools/clientcmd/api/v1"
-)
-
-func makeKubeconfig(endpoint, username, token, namespace string) *v1.Config {
-	clustername := strings.Replace(endpoint, ".", "-", -1)
-	authinfoname := username + "/" + clustername
-	contextname := namespace + "/" + clustername + "/" + username
-
-	return &v1.Config{
-		APIVersion: "v1",
-		Kind:       "Config",
-		Clusters: []v1.NamedCluster{
-			{
-				Name: clustername,
-				Cluster: v1.Cluster{
-					Server:                "https://" + endpoint,
-					InsecureSkipTLSVerify: true,
-				},
-			},
-		},
-		AuthInfos: []v1.NamedAuthInfo{
-			{
-				Name: authinfoname,
-				AuthInfo: v1.AuthInfo{
-					Token: token,
-				},
-			},
-		},
-		Contexts: []v1.NamedContext{
-			{
-				Name: contextname,
-				Context: v1.Context{
-					Cluster:   clustername,
-					Namespace: namespace,
-					AuthInfo:  authinfoname,
-				},
-			},
-		},
-		CurrentContext: contextname,
-	}
-}
